Multiple bug fixes.
Bugs:
** #208
** #209
