+++ includes/configs/Config.php
@@ -8,15 +8,6 @@
 namespace TooBasic;
 
 /**
- * @class ConfigException
- * This exeption is thrown whenever 'Config' or 'ConfigsManager' finds a halting
- * error.
- */
-class ConfigException extends Exception {
-	
-}
-
-/**
  * @class Config
  * This class represent a JSON file and its properties providing a simpler access
  * and a controlled and easier loading mechanism.
