+++ includes/exceptions/CacheException.php
@@ -1,0 +1,15 @@
+<?php
+
+/**
+ * @file CacheException.php
+ * @author Alejandro Dario Simi
+ */
+
+namespace TooBasic;
+
+/**
+ * @class CacheException
+ */
+class CacheException extends Exception {
+	
+}
