+++ includes/exceptions/ConfigException.php
@@ -1,0 +1,17 @@
+<?php
+
+/**
+ * @file ConfigException.php
+ * @author Alejandro Dario Simi
+ */
+
+namespace TooBasic;
+
+/**
+ * @class ConfigException
+ * This exeption is thrown whenever 'Config' or 'ConfigsManager' finds a halting
+ * error.
+ */
+class ConfigException extends Exception {
+	
+}
