+++ includes/exceptions/DBException.php
@@ -1,0 +1,15 @@
+<?php
+
+/**
+ * @file DBException.php
+ * @author Alejandro Dario Simi
+ */
+
+namespace TooBasic;
+
+/**
+ * @class DBException
+ */
+class DBException extends Exception {
+	
+}
