+++ includes/exceptions/DBStructureManagerException.php
@@ -1,0 +1,17 @@
+<?php
+
+/**
+ * @file DBStructureManagerException.php
+ * @author Alejandro Dario Simi
+ */
+
+namespace TooBasic\Managers;
+
+/**
+ * @class DBStructureManagerExeption
+ * This is a specific exception class for fatal errors found on a database
+ * structure check and update.
+ */
+class DBStructureManagerException extends \TooBasic\DBException {
+	
+}
