+++ includes/exceptions/Exception.php
@@ -46,17 +46,3 @@ class Exception extends \Exception {
 		include $Defaults[GC_DEFAULTS_EXCEPTION_PAGE];
 	}
 }
-
-/**
- * @class CacheException
- */
-class CacheException extends Exception {
-	
-}
-
-/**
- * @class DBException
- */
-class DBException extends Exception {
-	
-}
