+++ includes/exceptions/FieldFilterException.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * @file FieldFilterException.php
+ * @author Alejandro Dario Simi
+ */
+
+namespace TooBasic\Representations;
+
+/**
+ * @class FieldFilterException
+ * This exception represent any found error related to field filters.
+ */
+class FieldFilterException extends \TooBasic\Exception {
+	
+}
