+++ includes/exceptions/MagicPropException.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * @file MagicPropException.php
+ * @author Alejandro Dario Simi
+ */
+
+namespace TooBasic;
+
+/**
+ * @class MagicPropException
+ * This exeption is thrown whenever 'MagicProp' finds a halting error.
+ */
+class MagicPropException extends Exception {
+	
+}
