+++ includes/exceptions/RestManagerException.php
@@ -1,0 +1,15 @@
+<?php
+
+/**
+ * @file RestManagerException.php
+ * @author Alejandro Dario Simi
+ */
+
+namespace TooBasic\Managers;
+
+/**
+ * @class RestManagerException
+ */
+class RestManagerException extends Exception {
+	
+}
