+++ includes/exceptions/SApiReaderAbstractException.php
@@ -1,0 +1,15 @@
+<?php
+
+/**
+ * @file SApiReaderAbstractException.php
+ * @author Alejandro Dario Simi
+ */
+
+namespace TooBasic;
+
+/**
+ * @class SApiReaderAbstractException
+ */
+class SApiReaderAbstractException extends SApiReaderException {
+	
+}
