+++ includes/exceptions/SApiReaderException.php
@@ -1,0 +1,15 @@
+<?php
+
+/**
+ * @file SApiReaderException.php
+ * @author Alejandro Dario Simi
+ */
+
+namespace TooBasic;
+
+/**
+ * @class SApiReaderException
+ */
+class SApiReaderException extends \TooBasic\Exception {
+	
+}
