+++ includes/exceptions/SApiReportException.php
@@ -1,0 +1,17 @@
+<?php
+
+/**
+ * @file SApiReportException.php
+ * @author Alejandro Dario Simi
+ */
+
+namespace TooBasic;
+
+/**
+ * @class SAReporterException
+ * This type of exception is thrown whenever a error occurs with Simple API
+ * Reporter.
+ */
+class SApiReportException extends Exception {
+	
+}
